Initial version
A reversed library for supporting the A* TechAir airbag system
written in Rust. A cli is also provided that leverages the library
to allow the user to read sensors and modify firmware versions.

A GUI could be written perhaps in Gtk that would allow users on
other operating systems to use their airbag system.

Note that this is far from a clean implementation however
it sets the foundations for one by codifying what was RE'ed
from the Winblows tool that comes with the system.

Signed-off-by: Edward O'Callaghan <edward@antitrust.cc>